third_party/blink/web_tests/virtual/customizable-select-disabled/html5lib/generated/run-menuitem-element-data-expected.txt

CONSOLE WARNING: A menuitem tag was parsed inside of a <select> which was not inserted into the document. This is not valid HTML and the behavior may be changed in future versions of chrome.
../resources/menuitem-element.dat: PASS
